+++ cypress/e2e/productGroup_CRUD.cy.js
@@ -1,4 +1,4 @@
-import { product_group } from "../../PageObject/productGroupPOM.cy";
+import { product_group } from "../PageObject/productGroupPOM.cy";
 
 const prodGroup = new product_group
 const url = 'https://mts-bol-dev.inconstruction.website/product-management/group'
